+++ src/main/java/com/java_academy/logic/model/MessageObject.java
@@ -1,7 +1,5 @@
 package com.java_academy.logic.model;
 
-import com.java_academy.logic.Players;
-
 public class MessageObject {
 
     private Players player;
