+++ src/main/java/com/java_academy/logic/state_machine/NewGameState.java
@@ -1,7 +1,7 @@
 package com.java_academy.logic.state_machine;
 
-import com.java_academy.logic.Players;
 import com.java_academy.logic.model.MessageObject;
+import com.java_academy.logic.model.Players;
 import com.java_academy.logic.state_machine.core.GameState;
 
 import java.util.function.Consumer;
